requirements.txt
rich>=14.0.0
prompt_toolkit>=3.0.50
textual>=3.30
pycryptodome>=3.22
